+++ slinky/src/keep_sections.rs
@@ -1,4 +1,4 @@
-/* SPDX-FileCopyrightText: © 2024 decompals */
+/* SPDX-FileCopyrightText: © 2025 decompals */
 /* SPDX-License-Identifier: MIT */
 
 use std::collections::HashSet;
@@ -10,15 +10,11 @@ use serde::Deserialize;
 #[derive(Clone, Debug, Eq, PartialEq, Deserialize)]
 #[serde(deny_unknown_fields)]
 #[serde(untagged)]
+#[derive(Default)]
 pub enum KeepSections {
     #[serde(skip)]
+    #[default]
     Absent,
     All(bool),
     WhichOnes(HashSet<String>),
-}
-
-impl Default for KeepSections {
-    fn default() -> Self {
-        Self::Absent
-    }
 }